refactor: move overlay use cases to application
