+++ extension/src/app/application/overlay/file-view.test.ts
@@ -1,8 +1,8 @@
 import { describe, expect, it } from "vitest";
-import type { DiffV2, SemanticDiffResponse } from "../app/domain/diff/types";
-import { must } from "../app/presentation/util/must";
-import { createFileView, type FileResult } from "./fileView";
-import type { View } from "./toggle";
+import type { DiffV2, SemanticDiffResponse } from "../../domain/diff/types";
+import { must } from "../../presentation/util/must";
+import { createFileView, type FileResult } from "./file-view";
+import type { View } from "./view-mode";
 
 const DIFF: DiffV2 = { schema: "prefablens.diff.v2", unresolvedGuids: [], roots: [], loose: [] };
