+++ extension/src/app/application/overlay/view-mode.ts
@@ -1,0 +1,1 @@
+export type View = "raw" | "semantic";
